Add TelemetryLoader class and refactor SessionView
